+++ src/component/AvatarMenu/index.ts
@@ -1,0 +1,2 @@
+export * from "./AvatarMenu";
+export * from "./AvatarMenuWrapper";