+++ src/component/index.ts
@@ -1,4 +1,5 @@
 export * from "./Anchor";
+export * from "./AvatarMenu";
 export * from "./ErrorPage";
 export * from "./Link";
-export * from "./ThemeToggle";
+export * from "./Login";